requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.11.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==25.3.0
aurelio-sdk==0.0.19
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
blis==0.7.5
bokeh==3.4.3
build==1.4.0
cachetools==5.5.0
catalogue==2.0.6
certifi==2021.10.8
cffi==2.0.0
charset-normalizer==2.0.7
chromadb==1.4.1
click==8.0.3
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cryptography==46.0.3
cycler==0.11.0
cymem==2.0.6
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
et_xmlfile==2.0.0
exceptiongroup==1.2.1
executing==2.0.1
fastai==2.5.3
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.0
filelock==3.18.0
flatbuffers==25.12.19
fonttools==4.28.2
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google_search_results==2.4.2
googleapis-common-protos==1.66.0
grandalf==0.8
greenlet==3.2.3
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.3
importlib-metadata==6.11.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.3.2
kubernetes==35.0.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langcodes==3.3.0
langfuse==3.7.0
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
litellm==1.72.6
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.26.1
matplotlib==3.5.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.4
murmurhash==1.0.6
mypy_extensions==1.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
ollama==0.6.1
onnxruntime==1.19.2
openai==1.86.0
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.48b0
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parso==0.8.4
pathy==0.6.1
pdfminer.six==20251107
pdfplumber==0.11.8
Pillow==10.1.0
pinecone==7.0.2
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
posthog==5.4.0
preshed==3.0.6
prompt_toolkit==3.0.47
propcache==0.3.2
proto-plus==1.25.0
protobuf==5.29.5
psutil==6.0.0
pulsar-client==3.8.0
pure-eval==0.2.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.3
pycparser==2.23
pydantic==2.10.6
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.13.1
pyparsing==3.0.6
pypdf==5.6.0
PyPDF2==3.0.1
pypdfium2==5.3.0
PyPika==0.50.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.4
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.9
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.1
rpds-py==0.25.1
rsa==4.9
scikit-learn==1.0.1
scipy==1.13.1
semantic-router==0.1.8
serpapi==0.1.5
setuptools-scm==6.3.2
shellingham==1.5.4
six==1.16.0
smart-open==5.2.1
smmap==5.0.2
sniffio==1.3.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==2.0.41
srsly==2.4.2
stack-data==0.6.3
starlette==0.32.0.post1
streamlit==1.50.0
sympy==1.14.0
tenacity==8.5.0
thinc==8.0.13
threadpoolctl==3.0.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.25.0
validators==0.35.0
wasabi==0.8.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.14.0
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.20.1
zipp==3.19.2
zstandard==0.23.0
